Refactoring

+++ src/types.rs
@@ -8,12 +8,11 @@ pub type XYCoordinates = (i32, i32);
 pub type XYVector = (i32, i32);
 
 /// A (x, y) coordinates of a `FieldElement`.
-// TODO: usize じゃなくて u32 にする。
-pub type FieldElementPosition = (usize, usize);
+pub type FieldElementPosition = (u32, u32);
 
 /// A location of a `FieldObject`.
 /// 
 /// It consists of (x, y, id_of_field_object).  
 /// The (x, y) part is equal to a `FieldElementPosition`.
 // TODO: 他の type の型を使えないか。
-pub type FieldObjectLocation = (usize, usize, String);
+pub type FieldObjectLocation = (u32, u32, String);
